requirements.txt
pillow
kissfix
PyQt5==5.14.0
pyqtgraph